restricted memory for java
As described in the doc, the bbtools use more memory than indicated with the java flag. They recommend to set the mem as 0.85 of the total memory available. This is important for cluster submission.

the resources mem, will be used for the cluster and the resources.java_mem for the job.