Dev (#115)
* Update template_config.yaml

* made physcov optional.

The keyword `physcov`in pileup counts the bases between correctly mapped, not overlapping paired-end reads as covered. 
I think this is a great feature but it is incopatibl with asking bbmap to map to all ambigous sites `ambigous=all`

* filter on all possible mappings

Contigs should be filtered based on all possible mappings. 
As we have the option in the `align_reads_to_final_contigs` we should filter contigs based on all ambigous mapping results and not only on the best (default).

* merging less strict

* this parameters seems to me to merge more reads.

https://jgi.doe.gov/data-and-tools/bbtools/bb-tools-user-guide/bbmerge-guide/

* add trimming 20bp

* corecct codeword 'ambiguous'

* pileup cannot read sam file with ambigous alignement

* moves comments above the param it's describing; fixes typos

* renamed log files added QC subdirectoy. make it compatible with snakemake 5

* renamed all rules, fixed bug with QC report

* fixes #108 , standardisize the naming of log and benchmark files from assembly/logs to logs/assembly

* updated log file naming from logs to logs/annotate

* adapted megahit for snakemake 5, I had problems with the-- continue and d temp dir

* made qc_report work.

* removed 'clean' stage from first plot, wich is confusing

* corrected titles

* removed ssao

* added spades 3.12
merged reads are used sperately by spades.
added possibility to drop normalization

* corrected bugs- tested on spades and megahit

* added loose parameters
made filtering optional
not tested

* corrected bug with rel paths

* added qc report to output

* circular dependency on default values

* new version